+++ src/providers/helpers/clearInitiatedLogins.ts
@@ -1,5 +1,4 @@
 import { CrossWindowProvider } from 'lib/sdkWebWalletCrossWindowProvider';
-import { IframeProvider } from 'lib/sdkWebWalletIframeProvider';
 import { ProviderTypeEnum, ProviderType } from '../types/providerFactory.types';
 
 export const clearInitiatedLogins = (
@@ -18,15 +17,6 @@ export const clearInitiatedLogins = (
         if (crossWindowProvider.isInitialized()) {
           // Clean up the crossWindowProvider state without triggering a full logout.
           crossWindowProvider.onDestroy();
-        }
-        break;
-      }
-
-      case ProviderTypeEnum.metamask:
-      case ProviderTypeEnum.passkey: {
-        const iframeProvider = IframeProvider.getInstance();
-        if (iframeProvider.isInitialized()) {
-          iframeProvider.dispose();
         }
         break;
       }