+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[5.1.2](https://github.com/multiversx/mx-sdk-dapp/pull/1561)] - 2025-08-19
+
+- [Fixed `clearInitiatedLogins` breaks `IframeProvider` login](https://github.com/multiversx/mx-sdk-dapp/pull/1560)
+- [Rework websocket retry mechanism](https://github.com/multiversx/mx-sdk-dapp/pull/1559)
+- [Update sdk-wallet-connect-provider](https://github.com/multiversx/mx-sdk-dapp/pull/1558)
+- [Fixed issue with transaction toast title](https://github.com/multiversx/mx-sdk-dapp/pull/1555)
+- [Added pnpm](https://github.com/multiversx/mx-sdk-dapp/pull/1554)
+
 ## [[5.1.1](https://github.com/multiversx/mx-sdk-dapp/pull/1553)] - 2025-08-12
 
 - [Fixed sdk-core peerDependency options](https://github.com/multiversx/mx-sdk-dapp/pull/1552)